pyproject.toml
[tool.poetry]
name = "fatrate"
version = "0.1.0"
description = "Рейтинг жирдяев"
authors = ["MikhailTo <toshkin.mikhail@yandex.ru>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.15.0"
fluent-runtime = "^0.4.0"
pydantic-settings = "^2.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
